requirements.txt
absl-py==2.1.0
accelerate==1.6.0
datasets==3.4.1
evaluate==0.4.3
numpy==1.26.4
pillow==11.2.1
scikit-learn==1.5.1
scipy==1.15.2
setproctitle==1.3.5
tensorboard==2.19.0
torch==2.6.0
torch-geometric==2.6.1
tqdm==4.66.4
transformers==4.50.2
